refactor: restructure ledger facade into dir
